// packages/client/src/Redux/ActionTypes/CellActionTypes.ts
export enum ACTION_TYPES {
  ADD_CELL_AFTER = "add_cell_after",
  UPDATE_CELL = "update_cell",
  GET_CELLS = "get_cells",
  DELETE_CELLS = "delete_cell",
  MOVE_CELL = "move_cell",
  FETCH_CELLS_ERROR = "fetch_cells_error",
  FETCH_CELLS_COMPLETE = "fetch_cells_complete",
  FETCH_CELLS = "fetch_cells",
  SAVE_CELLS = "save_cells",
  SAVE_CELLS_ERROR = "save_cells_error",
}